fix(api): norm() returns Scalar (dtype-preserving), not double
Review of #1000 (ianmccul): returning a fixed double from norm() made
`x /= x.norm()` silently promote a Float tensor to Double through Div's
Tensor/scalar path. Return a Scalar carrying the input's own precision
instead (Float for Float/ComplexFloat, Double otherwise), so in-place
division stays dtype-preserving. Tensor::norm(), UniTensor::norm() and
both linalg::norm() overloads change double -> Scalar; the deprecated
Norm() forms are unchanged.

Python must not expose cytnx.Scalar, so the pybind norm bindings convert
to a native float (double(self.norm())) -- Python still gets a plain
float and derives type from dtype.

Internal callers (Block/BlockFermionic normalize_/Norm, Lanczos_ER/Gnd/
Gnd_Ut/Exp) keep their prior double behavior via an explicit double(...)
cast, so no algorithm's numerics change.

Also fix a pre-existing GCC build failure of the Python module: the
deprecated Tensor::Norm / UniTensor::Norm / UniTensor::combineBonds
bindings had a #pragma GCC diagnostic inside the .def() chain, which GCC
rejects ("'#pragma' is not allowed here"; only clang tolerated it). The
suppression now lives in file-scope helper functions, leaving the chain
pragma-free.

Tests: C++ norm/Norm/normalize/Lanczos 69/69 (incl. Float/ComplexFloat);
pytests 116 passed / 1 skipped (norm returns float; deprecated bindings
still warn); dtype-preservation assertions added.

+++ pybind/linalg_py.cpp
@@ -847,10 +847,12 @@ void linalg_binding(py::module &m) {
   #if defined(__GNUC__) || defined(__clang__)
     #pragma GCC diagnostic pop
   #endif
+  // norm() returns a Scalar in C++ (dtype-preserving); hand Python a native float so
+  // cytnx.Scalar is not exposed on the Python surface.
   m_linalg.def(
-    "norm", [](cytnx::UniTensor &T1) { return T1.norm(); }, py::arg("T1"));
+    "norm", [](cytnx::UniTensor &T1) { return double(T1.norm()); }, py::arg("T1"));
   m_linalg.def(
-    "norm", [](cytnx::Tensor &T1) { return T1.norm(); }, py::arg("T1"));
+    "norm", [](cytnx::Tensor &T1) { return double(T1.norm()); }, py::arg("T1"));
 
   m_linalg.def("Dot", &cytnx::linalg::Dot, py::arg("T1"), py::arg("T2"));
   m_linalg.def(
